+++ frontend/src/App.vue
@@ -1,19 +1,30 @@
 <script setup lang="ts">
-import { onMounted, watch } from 'vue'
+import { onMounted, ref, provide } from 'vue'
 import { RouterView } from 'vue-router'
 import HeaderComponent from './components/HeaderComponent.vue'
 import ThemeComponent from './components/ThemeComponent.vue'
 import { useUserStore } from './stores/user'
+import { useRoute } from 'vue-router'
+const route = useRoute()
 const userStore = useUserStore()
-onMounted(() => {
-  if (userStore) {
-    console.log(userStore)
+const userData = ref([])
+/* onMounted(() => { */
+if (userStore.userID) {
+  try {
+    const res = userStore
+      .getUser(Number(userStore.userID))
+      .then((response) => response.json())
+      .then((data) => (userData.value = data))
+    provide('userData', userData)
+  } catch (error) {
+    console.error(error)
   }
-})
+}
+console.log(userData)
 </script>
 <template>
   <div>
-    <HeaderComponent />
+    <HeaderComponent :userData="userData" />
     <!--should probaby put this on the home page; in app for now can we move this later when katherine is done or smth-->
     <ThemeComponent />
     <RouterView />
